Changed SDCC_Path.bat to Compile.bat and added make clean and make after the path is set.
Removed cygwin1.dll zip file as the problem encountered wasn't related to this file or the 8051 SDCC side of things at all.

Was in fact related to msys-1.0.dll in the WinAVR directory having issues
with make. Possible it was also related to cygwin1.dll as well for WinAVR, but nothing to do with SDCC. WinAVR uses mingw as the default make tools.